ページコメント削除機能を追加し、コメントリストの表示を改善しました。また、非同期処理を最適化しました。
