+++ next/src/app/[locale]/(common-layout)/user/[handle]/page/[slug]/_components/comment/_components/page-comment-list/page-comment-item/_db/mutations.server.ts
@@ -1,0 +1,7 @@
+import { prisma } from "@/lib/prisma";
+
+export async function deletePageComment(pageCommentId: number) {
+	return await prisma.pageComment.delete({
+		where: { id: pageCommentId },
+	});
+}
